Adding the EmailNotificationStatus domain model
This is used while the email notification is being
processed itself. It’s a slimmed down version
of the EmailNotificationDelivery domain model, as it
wouldn’t concern itself with anything except the
notification in transit. 

Once it’s completed it will redirect to the latest
record, where more details/logs can be inspected.